+++ packages/client/src/client.d.ts
@@ -11,6 +11,11 @@ declare module '*.png' {
   export default content;
 }
 
+declare module '*.jpg' {
+  const content: string;
+  export default content;
+}
+
 declare module '*.mp3' {
   const content: string;
   export default content;
