+++ packages/client/src/pages/Game/Game.scss
@@ -22,303 +22,3 @@
     font-size: 15px;
   }
 }
-
-.game-menu {
-  width: 200px;
-  display: flex;
-  flex-direction: column;
-  height: fit-content;
-  border-radius: 10px;
-  box-shadow: 0px 0px 13px 1px rgba(106, 139, 165, 0.58);
-  padding: 1rem;
-
-  &__header {
-    color: var(--color-messages);
-    width: 100%;
-    text-align: center;
-    font-size: 20px;
-  }
-
-  &__submenu {
-    list-style: none;
-    padding: 0;
-
-  }
-
-  &__link {
-    color: var(--color-messages);
-    text-decoration: none;
-    cursor: pointer;
-
-    &_color-red {
-      color: red;
-    }
-  }
-
-  &_theme_shark {
-    background-color: rgba(255, 255, 255, .1);
-    // font-family: 'PressStart', sans-serif;
-    // font-size: smaller;
-  }
-}
-
-.game-info {
-  display: flex;
-  flex-direction: column;
-  width: 200px;
-
-  &__next-figure {
-    width: 200px;
-    height: 200px;
-    border-radius: 10px;
-    box-shadow: 0px 0px 13px 1px rgba(106, 139, 165, 0.58);
-  }
-
-  &__user-info {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  &__avatar {
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-    cursor: pointer;
-  }
-
-  &__user-name {
-    text-decoration: none;
-    min-height: 50px;
-    cursor: pointer;
-    margin-top: 10px;
-  }
-
-  &__score {
-    padding-top: 20px;
-  }
-
-  &__canvas-figure {
-    padding-top: 15%;
-  }
-
-  &__sound-controls {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-  }
-
-  &_theme_shark {
-    background-color: rgba(255, 255, 255, .1);
-  }
-}
-
-.game-screen {
-  width: 300px;
-  height: 600px;
-  display: flex;
-  justify-content: center;
-
-  &__h3 {
-    text-align: center;
-  }
-
-  &__canvas {
-    height: 100%;
-    position: relative;
-  }
-
-  &__start-button {
-    width: 200px;
-    height: 40px;
-    border: none;
-    border-radius: 8px;
-    background-color: var(--color-links);
-    color: var(--white);
-    position: absolute;
-    top: 300px;
-
-  }
-
-  &__game-end {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    width: 233px;
-    height: 178px;
-    border-radius: 10px;
-    background-color: var(--white);
-    position: absolute;
-    top: 250px;
-    padding: 10px;
-    box-shadow: 0 0 6px 0 rgba(106, 139, 165, 0.58);
-  }
-
-  &__end-button,
-  &__error-button {
-    width: 200px;
-    height: 40px;
-    border: none;
-    border-radius: 8px;
-    background-color: var(--color-links);
-    color: var(--white);
-  }
-
-  &__error {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    width: 233px;
-    border-radius: 10px;
-    background-color: var(--white);
-    position: absolute;
-    top: 250px;
-    padding: 10px;
-    box-shadow: 0 0 6px 0 rgba(106, 139, 165, 0.58);
-  }
-}
-
-
-.game-over {
-  background: white;
-  height: 10rem;
-  width: 100%;
-  border-radius: 10px;
-  box-shadow: 0px 0px 13px 1px rgba(106, 139, 165, 0.58);
-  padding: 0rem 1rem;
-  display: none;
-}
-.show {
-  display: table-caption;
-}
-
-.background {
-  position: absolute;
-  z-index: -100;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-
-  &_theme_shark {
-    background-repeat: no-repeat;
-    background-size: cover;
-  }
-}
-
-.select-theme {
-  padding-bottom: 20px;
-
-  &__select {
-    background: none;
-    border: none;
-
-  }
-}
-
-.sound-controls {
-  width: 20px;
-  height: 20px;
-  background-color: var(--color-backgrnd-gray-dark-theme);
-  -webkit-mask-position-x: center;
-  -webkit-mask-position-y: center;
-  -webkit-mask-repeat: no-repeat;
-  mask-repeat: no-repeat;
-  mask-position: center;
-  cursor: pointer;
-  
-  &__sound {
-    -webkit-mask-image: url('./../../assets/volume-high-solid.svg');
-    mask-image: url('./../../assets/volume-high-solid.svg');
-  }
-
-  &__equalizer {
-    -webkit-mask-image: url('./../../assets/sliders-solid.svg');
-    mask-image: url('./../../assets/sliders-solid.svg');
-  }
-
-  &__music {
-    -webkit-mask-image: url('./../../assets/music-solid.svg');
-    mask-image: url('./../../assets/music-solid.svg');
-  }
-
-  &_vol_mute {
-    background-color: var(--color-midgray);
-  }
-}
-
-.levels__sound-controls {
-  margin-left: 15px;
-  width: 15px;
-  height: 15px;
-  background-color: var(--color-dark-theme-mid-gray);
-  -webkit-mask-position-x: center;
-  -webkit-mask-position-y: center;
-  -webkit-mask-repeat: no-repeat;
-  mask-repeat: no-repeat;
-  mask-position: center;
-  cursor: pointer;
-  
-  &_sound {
-    -webkit-mask-image: url('./../../assets/volume-high-solid.svg');
-    mask-image: url('./../../assets/volume-high-solid.svg');
-  }
-
-  &_music {
-    -webkit-mask-image: url('./../../assets/music-solid.svg');
-    mask-image: url('./../../assets/music-solid.svg');
-  }
-}
-
-.game-info__sound-controls_theme_shark,
-.game-info__p_theme_shark,
-.game-info__user-name_theme_shark{
-  width: 100%;
-  background-color: rgba(255, 255, 255, .4);
-}
-
-.game-menu_theme_shark,
-.game-info__next-figure_theme_shark {
-  background-color: rgba(255, 255, 255, .2);
-}
-
-.game-menu__link_accent {
-  color: #cb0000;
-}
-
-.sound-controls__input {
-  -webkit-appearance: slider-vertical;
-  appearance: slider-vertical;
-  writing-mode: bt-lr;
-  width: 50px;
-}
-
-.levels {
-  display: flex;
-  flex-direction: row;
-  background-color: var(--white);
-  position: absolute;
-  z-index: 1100;
-  width: max-content;
-  border-radius: 8px;
-  padding: 10px;
-
-  &__hide {
-    cursor: pointer;
-    transform: scaleX(1.3);
-    color: var(--color-backgrnd-gray-dark-theme);
-  }
-  
-}
-
-.levels-background {
-  position: fixed;
-  z-index: 1050;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, .4);
-}
