+++ packages/client/src/pages/Game/Game_info.scss
@@ -1,0 +1,159 @@
+.game-info {
+  display: flex;
+  flex-direction: column;
+  width: 200px;
+
+  &__next-figure {
+    width: 200px;
+    height: 200px;
+    border-radius: 10px;
+    box-shadow: 0px 0px 13px 1px rgba(106, 139, 165, 0.58);
+
+    &_theme_shark {
+      background-color: rgba(255, 255, 255, 0.2);
+    }
+  }
+
+  &__user-info {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+
+  &__avatar {
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    cursor: pointer;
+  }
+
+  &__user-name {
+    text-decoration: none;
+    min-height: 50px;
+    cursor: pointer;
+    margin-top: 10px;
+
+    &_theme_shark {
+      width: 100%;
+      background-color: rgba(255, 255, 255, 0.4);
+    }
+  }
+
+  &__score {
+    padding-top: 20px;
+  }
+
+  &__canvas-figure {
+    padding-top: 15%;
+  }
+
+  &__sound-controls {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+
+    &_theme_shark {
+      width: 100%;
+      background-color: rgba(255, 255, 255, 0.4);
+    }
+  }
+
+  &__p {
+    &_theme_shark {
+      width: 100%;
+      background-color: rgba(255, 255, 255, 0.4);
+    }
+  }
+
+  &_theme_shark {
+    background-color: rgba(255, 255, 255, 0.1);
+  }
+}
+
+.sound-controls {
+  width: 20px;
+  height: 20px;
+  background-color: var(--color-backgrnd-gray-dark-theme);
+  -webkit-mask-position-x: center;
+  -webkit-mask-position-y: center;
+  -webkit-mask-repeat: no-repeat;
+  mask-repeat: no-repeat;
+  mask-position: center;
+  cursor: pointer;
+
+  &__sound {
+    -webkit-mask-image: url('./../../assets/volume-high-solid.svg');
+    mask-image: url('./../../assets/volume-high-solid.svg');
+  }
+
+  &__equalizer {
+    -webkit-mask-image: url('./../../assets/sliders-solid.svg');
+    mask-image: url('./../../assets/sliders-solid.svg');
+  }
+
+  &__music {
+    -webkit-mask-image: url('./../../assets/music-solid.svg');
+    mask-image: url('./../../assets/music-solid.svg');
+  }
+
+  &_vol_mute {
+    background-color: var(--color-midgray);
+  }
+
+  &__input {
+    -webkit-appearance: slider-vertical;
+    appearance: slider-vertical;
+    writing-mode: bt-lr;
+    width: 50px;
+  }
+}
+
+.levels {
+  display: flex;
+  flex-direction: row;
+  background-color: var(--white);
+  position: absolute;
+  z-index: 1100;
+  width: max-content;
+  border-radius: 8px;
+  padding: 10px;
+
+  &__hide {
+    cursor: pointer;
+    transform: scaleX(1.3);
+    color: var(--color-backgrnd-gray-dark-theme);
+  }
+
+  &__sound-controls {
+    margin-left: 15px;
+    width: 15px;
+    height: 15px;
+    background-color: var(--color-dark-theme-mid-gray);
+    -webkit-mask-position-x: center;
+    -webkit-mask-position-y: center;
+    -webkit-mask-repeat: no-repeat;
+    mask-repeat: no-repeat;
+    mask-position: center;
+    cursor: pointer;
+
+    &_sound {
+      -webkit-mask-image: url('./../../assets/volume-high-solid.svg');
+      mask-image: url('./../../assets/volume-high-solid.svg');
+    }
+
+    &_music {
+      -webkit-mask-image: url('./../../assets/music-solid.svg');
+      mask-image: url('./../../assets/music-solid.svg');
+    }
+  }
+}
+
+.levels-background {
+  position: fixed;
+  z-index: 1050;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.4);
+}
