+++ packages/client/src/pages/Game/Game_menu.scss
@@ -1,0 +1,44 @@
+.game-menu {
+  width: 200px;
+  display: flex;
+  flex-direction: column;
+  height: fit-content;
+  border-radius: 10px;
+  box-shadow: 0px 0px 13px 1px rgba(106, 139, 165, 0.58);
+  padding: 1rem;
+
+  &__header {
+    color: var(--color-messages);
+    width: 100%;
+    text-align: center;
+    font-size: 20px;
+  }
+
+  &__submenu {
+    list-style: none;
+    padding: 0;
+  }
+
+  &__link {
+    color: var(--color-messages);
+    text-decoration: none;
+    cursor: pointer;
+
+    &_color-red {
+      color: var(--color-red);
+    }
+  }
+
+  &_theme_shark {
+    background-color: rgba(255, 255, 255, 0.2);
+  }
+}
+
+.select-theme {
+  padding-bottom: 20px;
+
+  &__select {
+    background: none;
+    border: none;
+  }
+}
