+++ packages/client/src/pages/Game/themes/classic/classic-theme.ts
@@ -1,0 +1,16 @@
+import end from './../classic/sounds/end.mp3';
+import line from './../classic/sounds/line.mp3';
+import position from './../classic/sounds/position.mp3';
+import start from './../classic/sounds/start.mp3';
+
+import tetris from './../classic/music/TECHNOTRIS.mp3';
+
+export const classicMusic = tetris;
+
+export const classicSounds = {
+  start: start,
+  end: end,
+  fall: position,
+  position: position,
+  line: line,
+};
