+++ packages/client/src/scss/_root.scss
@@ -42,4 +42,6 @@
     --body-text-align: #{$body-text-align};
   }
   --body-bg: #{$body-bg};
+  --color-red: #{$color-red-light};
+
 }
